Add invoke-unit-tests job to release pipeline
The release-pipeline.yml file now includes a new job named `invoke-unit-tests`. This job runs on `ubuntu-latest` and is conditioned to execute only if the `new-version` job succeeds and the `validVersion` output is `True`. It depends on the completion of both `new-build` and `new-version` jobs.

Environment variables `buildVersion`, `projectName`, and `projectType` are set using outputs from the `new-version` job. The job steps include checking out the repository and publishing the executable project using a custom GitHub action `g4-api/g4-github-actions/actions/test-dotnet-project@v1`, configured with parameters such as `build_version`, `project_name`, `project_type`, a filter for `TestCategory=ExportData`, and a path to the test settings file.

Additionally, the `dotnet build` command in the build step has been updated to include an extra line break.