Refactor test settings file handling
Introduce a new environment variable `testSettingsFileName` with a default value of 'Default.runsettings'. Update the `runSettingsFileName` variable in the PowerShell script to use this new environment variable. Simplify the `Save` method call for `$runSettings` by removing redundant double quotes. Add a conditional check to the "Invoke Unit Tests" step to ensure it only runs if the "Set the Default Test Settings File" step is successful. Update the `test-settings` parameter in the "Invoke Unit Tests" step to use the new environment variable.