Add job to publish test results and new env variable
Added a new environment variable `resultsArtifactName` to the `env`
section, which constructs the name of the unit test results artifact
using the build version.

Introduced a new job `publish-test-results` to the pipeline, which
runs on `ubuntu-latest` and is dependent on the `invoke-unit-tests`
job. The `publish-test-results` job includes the following steps:
- Checkout the repository.
- List downloadable artifacts using the GitHub API.
- Download the test results artifact using the `actions/download-artifact`
  action, with the name constructed from the `resultsArtifactName`
  environment variable.
- Publish the test results using the `EnricoMi/publish-unit-test-result-action`,
  specifying the check name and the path to the test result files.