Add step to collect and display coverage summary
Added a new step named "Collect Job Summary" to the `release-pipeline.yml` file.
This step sets the working directory to the value of `STAGE_DIRECTORY` environment variable.
It uses the PowerShell shell to execute the script.
The script reads the content of the coverage report file specified by `steps.new-coverage-file.outputs.coverageSummaryFile`.
The content of the coverage report is then appended to the GitHub Actions job summary, making it visible in the workflow run details.