Add "Collect Job Summary" step to release pipeline
A new step named "Collect Job Summary" has been added to the
`release-pipeline.yml` file. This step runs in the PowerShell
shell and operates within the directory specified by the
`STAGE_DIRECTORY` environment variable. It reads the content
of the summary file generated by the `new-unit-test-summary`
step and appends this content to the GitHub Actions job
summary. This enhancement ensures that the unit test summary
is included in the workflow run details, making it easier to
review the test results directly from the workflow.