Update report paths in release-pipeline.yml
Modified the paths for test and coverage reports in `release-pipeline.yml`. Instead of using `matrix.test-project` variables directly, the paths are now constructed by appending `matrix.test-project.report` and `matrix.test-project.coverage` to the `dropPath` variable, which is derived from `env.STAGE_DIRECTORY`. This change ensures that the report files are placed in the specified drop directory.