Add new-release job and update artifact naming in pipeline
Updated `release-pipeline.yml` to change the naming convention for NuGet packages artifact to `nuget-packages-drop-v${{ env.BUILD_VERSION }}`. Adjusted the path for uploading build artifacts for utility projects. Added a new job `new-release` to create a GitHub release and tag it with the build version, conditioned to run only if `NewVersion` job succeeds and the version is valid. The `new-release` job includes steps to checkout the repository, download build artifacts, and create a GitHub release with `.zip` and `.nupkg` files.