Reintroduce and update CI/CD jobs and conditions
Reintroduce the `InvokeTests` job with its original configuration, including steps for checking out the repository, setting test settings, invoking tests, generating summaries, and publishing artifacts and summaries.

Update the `publish-nuget-packages` job to include the build version in its name and re-enable the condition to run only if `InvokeTests` is successful.

Uncomment and update the `Publish Build Artifacts to NuGet Feed` step to use environment variables for the API key and source.

Update the `NewRelease` job to include additional conditions to run only if `InvokeTests` and `PublishTools` are successful.