Add new optional property to JSON object with default value
A new property has been introduced to the JSON object. This
property is not mandatory ("mandatory": false) and has a
default value of "Xpath". The description provided explains
that this property specifies how the automation framework
should identify or locate the target element within the UI,
defining the strategy or method for selecting the element on
which the assertion or action will be performed.