Update using directives and refactor test methods
Updated `using` directives to include `G4.Abstraction.WebDriver` and `System.Collections.Generic`. Renamed methods `GetStageDriver` to `GetStageDriverTest` and `GetStateDriver` to `GetStateDriverTest`. Commented out and renamed test method `A` to `ReadFromFileTest`, with slight refactoring for clarity. Added but commented out `NewDriverTest` method for creating and disposing of a new driver instance.